Iterate Ember arrays and forEach-able objects only when @ember/array loads
The default template iterator knew about Ember arrays and objects with a
`forEach` method. Both are Ember conventions, so `@ember/array` now
registers them as an iterator extension when it loads. The default
handles native arrays and native iterables.
